Update to the new peci_GetCPUID() call
peci_GetCPUID() was updated to return the CPU model and stepping
separately.  Only the model is required to determine the correct
PECI calls for the host error monitor, so the stepping can be
ignored.

Tested:
Injected an ERR2 timeout and confirmed that the CPU number can
be correctly identified using PECI.

Change-Id: I4d62e499cc8a0d776bc3d9959b2c67b1025f318c
Signed-off-by: Jason M. Bills <jason.m.bills@intel.com>